Dynamic table & relationship creation for models
All models inside the database/models directory will automatically be
added to the tables of the sqlite database upon creation. This is
achieved using importlib and pkgutil. This will ensure that upon
database creation all tables and relationships are created even
if there are no references to imports for the specific models.